Use Plug and Homebrew instead of port and Vundle
Moving towards a more minimal and light-weight setup with Plug and brew
instead of Vundle and Port.

Also moved the python setup code into its own script. This new script
file is not tested yet and is more meant as a way to keep the code
around until I decide if it's useful or not.